fix(storybook): unblock build and prevent style override leakage
The Angular Vite JIT plugin base64-encodes inline component styles
into virtual asset filenames. The repeated harness style block in each
story decoded to ~600 chars, exceeding the 255-byte filename limit and
breaking the build with ENAMETOOLONG. Extract the shared styles to
src/components/story-harness.css and reference via styleUrl.

Separately, the icon-variants and svg-icons stories each declared
their own STYLE_TAG_ID and only removed their own override tag, so
custom-property overrides (e.g. --tbx-mat-banner-icon-size: 3rem from
"Large Icons") leaked across files and into stories with no override
decorator. Centralize the helpers in src/components/story-overrides.ts
behind a single shared id and add a global cleanup decorator in
.storybook/preview.ts that strips the tag before every story renders.
