+++ .storybook/preview.ts
@@ -4,6 +4,7 @@ import { provideAnimationsAsync } from '@angular/platform-browser/animations/asy
 import { MAT_ICON_DEFAULT_OPTIONS } from '@angular/material/icon';
 import { TBX_MAT_BANNER_PROVIDER_CONFIG } from '../src/tokens/banner-provider-config.token';
 import { TbxMatBannerSeverityFontIconService } from '../src/services/banner-severity-font-icon.service';
+import { removeStoryOverrideStyleTag } from '../src/components/story-overrides';
 
 // M3 prebuilt theme — provides typography, shape, and state-layer tokens.
 import '@angular/material/prebuilt-themes/azure-blue.css';
@@ -12,6 +13,11 @@ import '../src/styles/_tbx-mat-banners.scss';
 
 const preview: Preview = {
     decorators: [
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        (story: () => any) => {
+            removeStoryOverrideStyleTag();
+            return story();
+        },
         applicationConfig({
             providers: [
                 provideAnimationsAsync(),
