+++ src/components/banner-svg-icons.stories.ts
@@ -7,28 +7,9 @@ import { MAT_ICON_DEFAULT_OPTIONS } from '@angular/material/icon';
 import { TbxMatBannerService } from '../services/banner.service';
 import { TbxMatBannerSeveritySvgIconService } from '../services/banner-severity-svg-icon.service';
 import { TBX_MAT_BANNER_PROVIDER_CONFIG } from '../tokens/banner-provider-config.token';
+import { withCustomProperties, withDefaultProperties } from './story-overrides';
 
 // ─── CSS Custom Property Overrides ───────────────────────────────────────────
-
-const STYLE_TAG_ID = 'tbx-banner-svg-story-overrides';
-
-function withCustomProperties(css: string) {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    return (story: () => any) => {
-        document.getElementById(STYLE_TAG_ID)?.remove();
-        if (css) {
-            const style = document.createElement('style');
-            style.id = STYLE_TAG_ID;
-            style.textContent = css;
-            document.head.appendChild(style);
-        }
-        return story();
-    };
-}
-
-function withDefaultProperties() {
-    return withCustomProperties('');
-}
 
 const LARGE_ICON_CSS = `
     html {
@@ -81,32 +62,7 @@ function withSvgIcons() {
             <p class="state">Active: {{ banner.isActive() }} &middot; Pending: {{ banner.pendingCount() }}</p>
         </div>
     `,
-    styles: `
-        .harness {
-            font-family: Roboto, sans-serif;
-            padding: 1.5rem;
-        }
-
-        h3 {
-            margin: 1.5rem 0 0.5rem;
-        }
-
-        h3:first-of-type {
-            margin-top: 0;
-        }
-
-        .button-group {
-            display: flex;
-            flex-wrap: wrap;
-            gap: 0.5rem;
-        }
-
-        .state {
-            margin-top: 1rem;
-            font-size: 0.875rem;
-            color: #666;
-        }
-    `,
+    styleUrl: './story-harness.css',
 })
 class BannerSvgIconsHarnessComponent {
     readonly banner = inject(TbxMatBannerService);
