+++ src/components/story-harness.css
@@ -1,0 +1,56 @@
+.harness {
+    font-family: Roboto, sans-serif;
+    padding: 1.5rem;
+}
+
+h3 {
+    margin: 1.5rem 0 0.5rem;
+}
+
+h3:first-of-type {
+    margin-top: 0;
+}
+
+.button-group {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.5rem;
+}
+
+.state {
+    margin-top: 1rem;
+    font-size: 0.875rem;
+    color: #666;
+}
+
+.theme-note {
+    font-size: 0.8125rem;
+    color: #888;
+    border-left: 3px solid #ddd;
+    padding: 0.25rem 0.75rem;
+    margin: 0 0 1rem;
+}
+
+.result-panel {
+    margin-top: 1rem;
+    background: #f0f4ff;
+    border-left: 3px solid #1565c0;
+    padding: 0.5rem 0.75rem;
+}
+
+.result-panel pre {
+    font-size: 0.8125rem;
+    margin: 0.25rem 0 0;
+    white-space: pre-wrap;
+}
+
+.severity-stack {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
+
+.dismissed-note {
+    font-size: 0.875rem;
+    color: #666;
+}
